+++ components/case-studies-card/case-studies-card.module.scss
@@ -28,6 +28,7 @@
 
   @include breakpoint(medium-only) {
     max-width: 100%;
+    height: 21.4375rem;
     width: 100%;
     border-radius: 8px;
   }
