+++ components/case-studies-card/index.tsx
@@ -1,0 +1,3 @@
+import CaseStudiesCard from './case-studies-card';
+
+export default CaseStudiesCard;
