+++ components/case-studies/case-studies.module.scss
@@ -3,7 +3,7 @@
 
 .container {
   background-color: $color-bg-grey;
-  padding-top: 56px;
+  padding-top: 3.5rem;
 
   display: flex;
   flex-direction: column;
@@ -16,9 +16,9 @@
 .sectionTitle {
   @include h1;
 
-  margin-bottom: 8rem;
+  margin-bottom: 5rem;
   text-align: center;
-  padding: 91px 0 0;
+  padding: 5.5875 0 0;
 
   @include breakpoint(small) {
     margin-bottom: 2.5rem;
@@ -29,7 +29,7 @@
 .cards {
   display: flex;
   justify-content: center;
-  margin: 0 50px;
+  margin: 0 3.125rem;
 
   @include breakpoint(small) {
     flex-direction: column;
