+++ components/case-studies/content.tsx
