+++ components/case-studies/index.tsx
@@ -1,0 +1,3 @@
+import CaseStudies from './case-studies';
+
+export default CaseStudies;
