--- components/use-cases-card/index.tsx
@@ -1,3 +1,0 @@
-import UseCasesCard from './use-cases-card';
-
-export default UseCasesCard;
