--- components/use-cases/index.tsx
@@ -1,3 +1,0 @@
-import UseCases from './use-cases';
-
-export default UseCases;
