+++ pages/index.tsx
@@ -2,7 +2,7 @@ import Head from 'next/head';
 import WhyKedro from '../components/why-kedro';
 import Hero from '../components/hero';
 import Features from '../components/features';
-import UseCases from '../components/use-cases';
+import CaseStudies from '../components/case-studies';
 
 const Home = () => {
   return (
@@ -13,7 +13,7 @@ const Home = () => {
       <Hero />
       <WhyKedro />
       <Features />
-      <UseCases />
+      <CaseStudies />
     </>
   );
 };
